ch12/requirements.txt
langgraph
langchain
langchain_openai
langchain_community
chromadb
aiohttp
tiktoken
beautifulsoup4
html2text
python-dotenv 
langgraph-supervisor
fastmcp